+++ test/src/reflection/user_with_reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.8
+// BUILDER: reflection_factory/1.2.9
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -112,7 +112,7 @@ class TestAddressWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestAddressWithReflection$reflection withObject(
@@ -360,7 +360,7 @@ class TestCompanyWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestCompanyWithReflection$reflection withObject(
@@ -668,7 +668,7 @@ class TestDataWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestDataWithReflection$reflection withObject([TestDataWithReflection? obj]) =>
@@ -903,7 +903,7 @@ class TestDomainWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestDomainWithReflection$reflection withObject(
@@ -1255,7 +1255,7 @@ class TestEnumWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestEnumWithReflection$reflection withObject([TestEnumWithReflection? obj]) =>
@@ -1325,7 +1325,7 @@ class TestOpAWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestOpAWithReflection$reflection withObject([TestOpAWithReflection? obj]) =>
@@ -1554,7 +1554,7 @@ class TestOpBWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestOpBWithReflection$reflection withObject([TestOpBWithReflection? obj]) =>
@@ -1797,7 +1797,7 @@ class TestOpWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestOpWithReflection$reflection withObject([TestOpWithReflection? obj]) =>
@@ -2047,7 +2047,7 @@ class TestTransactionWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestTransactionWithReflection$reflection withObject(
@@ -2262,7 +2262,7 @@ class TestUserWithReflection$reflection
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.8');
+  Version get reflectionFactoryVersion => Version.parse('1.2.9');
 
   @override
   TestUserWithReflection$reflection withObject([TestUserWithReflection? obj]) =>